decouple types to avoid circular imports
